separate layout wrapper for wallet app
