+++ packages/suite-desktop/pages/wallet/account/transactions/index.tsx
@@ -1,0 +1,3 @@
+import Account from '@wallet-views/account/send';
+
+export default Account;
