+++ packages/suite/src/components/wallet/TopNavigation/index.messages.ts
