+++ packages/suite/src/views/wallet/account/send/index.tsx
@@ -5,6 +5,7 @@ import { Text } from 'react-native';
 import { Button } from '@trezor/components';
 import { State } from '@suite-types/index';
 import { goto } from '@suite-actions/routerActions';
+import Layout from '@wallet-components/Layout';
 
 interface Props {
     suite: State['suite'];
@@ -15,7 +16,7 @@ const Wallet = (props: Props) => {
     const { pathname, params } = props.router;
     const baseUrl = `${pathname}#/${params.coin}/`;
     return (
-        <>
+        <Layout>
             <Text>
                 {params.coin} Account {params.accountId} Send Page
             </Text>
@@ -39,7 +40,7 @@ const Wallet = (props: Props) => {
             <Button variant="success" onClick={() => goto('/wallet/account/sign-verify', true)}>
                 Sign verify
             </Button>
-        </>
+        </Layout>
     );
 };
