+++ packages/suite/src/views/wallet/account/sign-verify/index.tsx
@@ -6,7 +6,7 @@ import { Button } from '@trezor/components';
 import { State } from '@suite-types/index';
 import { goto } from '@suite-actions/routerActions';
 import Title from '@suite/components/wallet/Title';
-
+import Layout from '@wallet-components/Layout';
 interface Props {
     suite: State['suite'];
     router: State['router'];
@@ -16,7 +16,7 @@ const Wallet = (props: Props) => {
     const { pathname, params } = props.router;
     const baseUrl = `${pathname}#/${params.coin}/`;
     return (
-        <>
+        <Layout>
             <Title>
                 {params.coin} Account {params.accountId} Sign-Verify Page
             </Title>
@@ -40,7 +40,7 @@ const Wallet = (props: Props) => {
             <Button variant="success" onClick={() => goto('/wallet/account/sign-verify', true)}>
                 Sign verify
             </Button>
-        </>
+        </Layout>
     );
 };
